Added test mocks data using j-unit 5

+++ src/main/kotlin/com/vecter/tutorial/springboot/vecter/model/Bank.kt
@@ -2,6 +2,6 @@ package com.vecter.tutorial.springboot.vecter.model
 
 data class Bank(
         val accountNumber: String,
-        val trust : Double,
-        val transactionFee : Int
+        val trust: Int,
+        val transactionFee: Int
 )
